requirements.txt
PyMuPDF>=1.19.0
ebooklib>=0.17.1
python-docx>=0.8.11
markdown>=3.3.4
pillow>=8.2.0
beautifulsoup4>=4.9.3
lxml>=4.6.3
tqdm>=4.61.0
py-kindlegen>=0.3.1
click>=8.0.1
pytest>=6.2.5
black>=21.5b2
isort>=5.9.1
flake8>=3.9.2
mypy>=0.812
